changed percent column to a float to accomodate percents of format XX.X%

+++ CreateTables.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS ProjStateTax(
     StateID CHAR(2) NOT NULL,
     MinIncome INT NOT NULL,
     MaxIncome INT NOT NULL,
-    Percent SMALLINT NOT NULL,
+    Percent FLOAT NOT NULL,
     PRIMARY KEY (StateBracketID)
 );
 
@@ -38,7 +38,7 @@ CREATE TABLE IF NOT EXISTS ProjSplit(
     SplitID INT NOT NULL AUTO_INCREMENT,
     BudgetID INT NOT NULL,
     Name VARCHAR(40),
-    Percent SMALLINT NOT NULL,
+    Percent FLOAT NOT NULL,
     FOREIGN KEY(BudgetID) REFERENCES ProjBudget(BudgetID),
     PRIMARY KEY(SplitID)
 );
